Don't try to double delete jobs with :delete_first.  Removes large numbers of Beanstalk::NotFound errors for delete_first jobs

+++ lib/async_observer/worker.rb
@@ -208,12 +208,14 @@ class AsyncObserver::Worker
     f.call(job) if f
     job.delete if job.ybody[:delete_first]
     run_code(job)
-    job.delete()
+    job.delete() unless jobs.ybody[:delete_first]
   rescue ActiveRecord::RecordNotFound => ex
-    if job.age > 60
-      job.delete() # it's old; this error is most likely permanent
-    else
-      job.decay() # it could be replication delay so retry quietly
+    unless jobs.ybody[:delete_first]
+      if job.age > 60
+        job.delete() # it's old; this error is most likely permanent
+      else
+        job.decay() # it could be replication delay so retry quietly
+      end
     end
   end
